Move deprecated config types within deprecated client namespace
